### 2020-11-24 #2860 @whedon
[PRE REVIEW]: sntools: Event generator for supernova burst neutrinos
**Submitting author:** @JostMigenda (<a href="http://orcid.org/0000-0002-5350-8049">Jost Migenda</a>)
**Repository:** <a href="https://github.com/JostMigenda/sntools/" target ="_blank">https://github.com/JostMigenda/sntools/</a>
**Version:** v0.7
