### 2020-11-24 #2860 @JostMigenda
I can think of a few potential reviewers, but most of them are co-authors on [this recent paper](https://arxiv.org/abs/2011.00035), which I have also contributed to; so they may be excluded by the JOSS [Conflict of Interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy).
Looking at [the list](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822), I think bradkav and kostunin would be good candidates; perhaps also arjunsavel, olebole or benjaminrose.
