### 2020-11-26 #2860 @xuanxu
@HeloiseS would you be willing to review this paper? I think your interest in SuperNovae and knowledge of Python makes you a great fit for this submission.
