### 2020-11-26 #2860 @JostMigenda
Quick note: @HeloiseS and I were PhD students in the same department until 1.5 years ago and still interact socially (but have never collaborated on a paper or similar). I think that’s not covered by the [CoI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) (which requires at least 1 year since being at the same institution, and does not limit Twitter interactions), but I still wanted to mention it in the interest of transparency.

That said, I agree Héloïse would be a great fit for this topic and I am certain that she will be able to judge it impartially, so I’d be happy to have her as a reviewer.
