### 2020-12-01 #2860 @xuanxu
@benjaminrose would you be willing to review this paper? I think your interest in SuperNovae and knowledge of Python makes you a great fit for this submission.
