### 2020-12-02 #2860 @benjaminrose
@xuanxu, yes, I can review this project. I am currently reviewing another JOSS paper, so I can't get to this one for about a week.

I wanted to note that the repo was started by Chris Walker (a co-worker of mine) but it appears that he was only the author of this initial commit and not apart of the project. If I am mistaken, @JostMigenda, can you correct me?
