### 2020-12-02 #2860 @xuanxu
@benjaminrose great, thanks!

>  I can't get to this one for about a week.

No problem!

> the repo was started by Chris Walter (a co-worker of mine)

He's not in the [paper](https://github.com/openjournals/joss-papers/blob/joss.02860/joss.02860/10.21105.joss.02860.pdf)'s authors list and can't be found as [contributor to the code](https://github.com/openjournals/joss-reviews/issues/2860#issuecomment-732756040), so currently I don't see any COI.
Anyway, let's wait for confirmation of his (not) involvement from @JostMigenda
