### 2020-12-02 #2860 @JostMigenda
I confirm that Chris Walter was not involved.

(The reason he appears in the commit history is because we originally started out with [a pre-existing script from a different repo](https://github.com/WCSim/WCSim/commits/develop/sample-root-scripts/MakeKin.py), so I included the git history of that file. Chris just happened to start that different repo; as far as I can reconstruct, his [commit in sntools](https://github.com/JostMigenda/sntools/commit/5800c46ec8e75b3a438f4cb153b460c83946e345) corresponds to [this commit in the original repo](https://github.com/WCSim/WCSim/commit/24c724cd2cef2cfc6e525ec159f2b812f5f1a6fd).)
