### 2020-12-09 #2877 @benjaminrose
SNtools review

This looks really useful! I see a few things that need a bit of an adjustment before I can recommend it to be accepted in JOSS. I know this list looks long, but I promise they are not too complicated and we should be able to converge on a acceptable submission quickly.

- [ ] `sntools` is a command line app. That needs to be more clear at the top to the documentation and readme.md

- [ ] Where is a readable version of the documentation? All I found was the latex source. There are ways to get Traves CI or GitHub actions to compile the source. This can be saved to its own branch if desired. Please link to a user facing documentation in either the Readme.md, the GitHub repo metadata, or both.

- [ ] Why does the first example command not work without optional arguments? I feel like you should show an example that can easily take simple tweaks.

- [ ] Why is this example passing the default to detector?

- [ ] This first example takes a while to run. Is there a shorter simulation that can be used just to test the installation process? Also, I had no idea how long this simulation was going to take, nor how far along it was while running. When running someone else's input file, it would be nice to get some output that says what it will do before it does it. Maybe add this to the verbose mode and add that to the first example?

- [ ] In the paper: "It is an event rate calculator—not an event generator—and states in its own documentation that it is “not intended to replace full detector simulations.” In contrast, sntools is intended for use in conjunction with a full detector simulation to perform more advanced, in-depth studies." -- in conjunction with is not the same as in place of. What exactly does `sntools` add to supplement to detector simulations that SNOwGLoBES can't? Is it the event level, as opposed to rate, information?

- [ ] Is there a walk through/tutorial for writing an input file? This is important since writing your own input file is the main thing a researcher will do.

- [ ] Tests don't install the package via pip. They currently test the source code, good. But don't test what will be sent installed from pypi, less good. This explains how you are testing your project directory and not how it will be installed by users, https://hynek.me/articles/testing-packaging/. This is a valid test suite, but may not be doing what you expect.
