### 2020-12-10 JostMigenda/sntools#16 @JostMigenda
Documentation improvements
Various minor documentation improvements, as suggested by @benjaminrose in [this post](https://github.com/openjournals/joss-reviews/issues/2877#issuecomment-742051700) in the JOSS review.

- [x] Make clear(er) that sntools is a command line app.
- [x] Compiled version of documentation.
- [x] First example should run more quickly.
- [x] Simplify arguments of first example
