### 2020-12-10 JostMigenda/sntools#16 @benjaminrose
@JostMigenda, I like the current example becuase it shows many `sntools` options. I would keep it along side an improved first run experience.
