### 2020-12-10 36044f7
Explain in README where to find full documentation
