### 2020-12-10 2856ade
Explain in README where to find full documentation
