### 2020-12-10 JostMigenda/sntools#16 @JostMigenda
@benjaminrose Perhaps it’s best to have two separate examples—an introductory one (that’s very quick and uses few optional arguments for clarity), and a more advanced example (that can run for longer and demonstrate many common arguments)?

E.g. `sntools fluxes/intp2001.data --format nakazato --endtime 50` as an introduction (that runs in ~15 s on my machine) and `sntools fluxes/intp1301.data --format nakazato --ordering normal --distance 20 --detector SuperK --verbose`
