### 2020-12-10 JostMigenda/sntools#16 @JostMigenda
Regarding compiled version of the LaTeX documentation:
I’ve uploaded a pdf file to release v0.7 and [have added an action](https://github.com/JostMigenda/sntools/commit/8ffcb516ae57f7996f7018ad535f48db1e963be1) that should automatically generate a pdf file for every new release and attach it.
