### 2020-12-10 JostMigenda/sntools#16 @benjaminrose
Sounds great!
