### 2020-12-11 6bcfecb
Improved usage examples
