+++ doc/documentation.tex
@@ -64,12 +64,19 @@ Then, run \texttt{sntools --version} to check whether it’s working.
 
 \subsection{Quick Start Guide}
 
-An example usage of sntools may look as follows:
+To generate your first supernova neutrino events with sntools, open a terminal window and type in the following command:
 
-\texttt{sntools fluxes/intp2001.data --format nakazato --ordering normal --detector HyperK}
+\texttt{sntools fluxes/intp2001.data --format nakazato --endtime 50}
 
-This generates events for a supernova in Hyper-Kamiokande, assuming normal mass ordering and the neutrino flux given in the input file \texttt{fluxes/intp2001.data}\footnote{This sample file, which is for a \SI{20}{M_\odot} progenitor from~\cite{Nakazato2013}, is included as part of sntools. See \url{http://asphwww.ph.noda.tus.ac.jp/snn/} for more information.}, which is in the \texttt{nakazato} format.
+This generates events for the first \SI{50}{ms} of a supernova based on the neutrino flux given in the input file \texttt{fluxes/intp2001.data}\footnote{This sample file, which is for a \SI{20}{M_\odot} progenitor from~\cite{Nakazato2013}, is included as part of sntools. See \url{http://asphwww.ph.noda.tus.ac.jp/snn/} for more information.} (which is in the \texttt{nakazato} format) and writes them to a file named \texttt{outfile.kin} in the current directory.
 % TODO: This file is not currently included in the pip-based install.
+
+A more realistic usage that demonstrates more of sntools’ capabilities looks like this:
+
+\texttt{sntools fluxes/intp2001.data --format nakazato --channel es --detector SuperK --distance 20 --verbose --output intp2001es.kin}
+
+This uses the neutrino flux given in the same input file to generate neutrino-electron elastic scattering events in Super-Kamiokande for a supernova at \SI{20}{kpc} distance. It also produces more verbose output, which lets you see that sntools generates events separately for \nue, \nuebar and \nux (which have different fluxes and cross sections), before merging them into an output file named \texttt{intp2001es.kin}.
+
 
 The input file is a required argument. The following optional arguments are available:
 
