### 2020-12-11 48a9614
explicitly say that sntools is a command line app as suggested by @benjaminrose in the JOSS review
