### 2020-12-11 5871a9e
explicitly say that sntools is a command line app as suggested by @benjaminrose in the JOSS review
