### 2020-12-11 7c605ee
minor documentation fixes
