+++ doc/documentation.tex
@@ -75,7 +75,8 @@ A more realistic usage that demonstrates more of sntools’ capabilities looks l
 
 \texttt{sntools fluxes/intp2001.data --format nakazato --channel es --detector SuperK --distance 20 --verbose --output intp2001es.kin}
 
-This uses the neutrino flux given in the same input file to generate neutrino-electron elastic scattering events in Super-Kamiokande for a supernova at \SI{20}{kpc} distance. It also produces more verbose output, which lets you see that sntools generates events separately for \nue, \nuebar and \nux (which have different fluxes and cross sections), before merging them into an output file named \texttt{intp2001es.kin}.
+This uses the neutrino flux given in the same input file to generate neutrino-electron elastic scattering events in Super-Kamiokande for a supernova at \SI{20}{kpc} distance.
+It also produces more verbose output, which lets you see that sntools generates events separately for \nue, \nuebar, \nux and \nuxbar (which have different fluxes and cross sections), before merging them into an output file named \texttt{intp2001es.kin}.
 
 
 The input file is a required argument. The following optional arguments are available:
