### 2020-12-11 JostMigenda/sntools#16 @JostMigenda
I’ve [updated the examples](https://github.com/JostMigenda/sntools/pull/18/commits/5d43ac10f4cf565d622e52ab796d692a186b1f40) as discussed, though I decided to use `--channel es` instead of `--ordering normal` for the second one: That makes the output shorter and easier to explain because it’s only simulating 4 channel/flavour combinations (instead of 14 for normal mass ordering and all channels).

Also [noted explicitly](https://github.com/JostMigenda/sntools/pull/18/commits/48a9614675d56bcf6a8f717a163c7851bd2bdc49) that it’s a command line app.

After those commits, I think #18 covers all the requested documentation improvements and I will merge it soon if there are no additional comments.
