### 2020-12-11 JostMigenda/sntools#19 @JostMigenda
Packaging improvements
Prompted by @benjaminrose’s suggestions in [this comment](https://github.com/openjournals/joss-reviews/issues/2877#issuecomment-742051700) on the JOSS review:

> Tests don't install the package via pip. They currently test the source code, good. But don't test what will be sent installed from pypi, less good. This explains how you are testing your project directory and not how it will be installed by users, https://hynek.me/articles/testing-packaging/. This is a valid test suite, but may not be doing what you expect.

Some background:
Since I didn’t have experience with Python packaging before working on sntools, I tried to stick close to the official recommendations ([1](https://packaging.python.org/guides/distributing-packages-using-setuptools/#your-package), [2](https://packaging.python.org/tutorials/packaging-projects/#creating-the-package-files)), which did not mention the `src` folder structure recommended in the linked blog post. (Though now that I’ve looked into it, it seems to me that this is [hotly debated](https://github.com/pypa/packaging.python.org/issues/320).)

I think my first step to improve packaging should be to set up a GitHub Action that automatically packages sntools and publishes it to PyPI whenever there’s a new release. (That will largely eliminate human error, which is probably the main source of packaging issues.) I’ll check how it’s possible to install & test as part of that action and will update this issue accordingly.
