### 2020-12-11 #2877 @JostMigenda
Many thanks for these comments, @benjaminrose!

The first 5 ones are discussed in JostMigenda/sntools#16.

> What exactly does sntools add to supplement to detector simulations that SNOwGLoBES can't? Is it the event level, as opposed to rate, information?

Exactly as you say—sntools generates individual events, whereas SNOwGLoBES only calculates the expected number of events in each energy bin.

> Is there a walk through/tutorial for writing an input file? This is important since writing your own input file is the main thing a researcher will do.

I don’t think most users of sntools will write their own input files. Instead, input files will typically be generated by the groups that perform computer simulations of SNe and make the resulting outputs (neutrino fluxes, gravitational wave signals, …) available to the community; while typical users of sntools work on a specific neutrino experiment and take input files from an available simulation to generate event files with sntools and then run those events through their experiment’s own detector simulation & event reconstruction toolchain.

If a simulations group makes their neutrino fluxes available in a format that’s not yet supported by sntools, an sntools user may need to add support for that new input format. That’s why I’ve included brief instructions for how to do that in chapter 3 of the documentation.

> Tests don't install the package via pip. They currently test the source code, good. But don't test what will be sent installed from pypi, less good.

I’ve opened JostMigenda/sntools#19 for this.
