### 2020-12-11 JostMigenda/sntools#19 @benjaminrose
Yes, this is a hotly debated topic. As such it should not hold up the acceptance to JOSS, but it is something to understand and document. Code always does what you say rather than what you intend.
