### 2020-12-22 #2877 @bradkav
Hi @JostMigenda, this looks like a really useful tool! The install proceeds smoothly and the examples look like they work pretty straightforwardly. I just have a few comments (most of which are simple/cosmetic) which will hopefully improve the accessibility for new users. My comments below are based on branch https://github.com/JostMigenda/sntools/tree/JostMigenda/issue16, which I think has already incorporated some changes suggested by @benjaminrose. (And sorry to drop this review on you just before Christmas.)

- [ ] **Examples:** After installing via pip, the files in the `fluxes/` folder are not locally available, so the command line examples don't work (that is, the user would first have to download the flux examples manually). I would suggest adding an extra line of script in the README, such as

	```bash
	curl -L https://raw.githubusercontent.com/JostMigenda/sntools/master/fluxes/intp2001.data -o intp2001.data
	```
  which will then allow the user to run the examples in the README immediately after the pip install.

- [ ] **Documentation:** In the README you mention a number of input/output file formats (Princeton, NUANCE, RATPAC). I think it would be useful to have some links in the README to point the user to some specifications of these file formats. As pointed out by @benjaminrose, there should also be a link to the full documentation somewhere in the README.

- [ ] **Tests:** At the moment, it doesn't look like there are any tests that the user can perform to check that everything is working correctly (if they've installed via pip). One possibility is that you could include the expected output of the command line examples in the README, so that the user can manually check (e.g. using a specified random seed to guarantee reproducibility). Another would be automated tests which the user can run after installing.

- [ ] **Community guidelines:** I can't see any information in the README about suggestions for how to contribute or how to get help. It looks like the latex documentation has some of this information, so once there's a clear link to a readable version of this, I think this can be marked as resolved.
