### 2021-01-05 565b3b0
Improved first-run experience Add sntools.setup() function to download sample input file and perform integration test.
