### 2021-01-05 #2877 @JostMigenda
Thanks for the comments @bradkav!

I spent the holiday period thinking about the first-run experience of new users; the result is in [JostMigenda/sntools#20](https://github.com/JostMigenda/sntools/pull/20) and should cover your bullet points on examples and tests.

Regarding documentation: [I’ve added a note](https://github.com/JostMigenda/sntools/commit/2856adeafc71ee63b9f9965a66ba865b7cfd3792#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R7) at the beginning of the README file explaining where to find full documentation. The full documentation contains more details and/or links to official documentation for the input/output file formats. Let me know if there’s anything specific I’m missing!

Regarding community guidelines: [I’ve added](https://github.com/JostMigenda/sntools/pull/21) a brief section on issue reporting and contributing to the README, with an explicit pointer to the detailed instructions in the full documentation.
