### 2021-01-08 #2877 @benjaminrose
@JostMigenda, I think this is really close. I see a few final tweaks that I think are needed before acceptance.

- [ ] Can you add a description of the available support options to the documentation? If it is just community (aka not lead by you) support centered around the GitHub issues, that is reasonable.
- [ ] I agree with @bradkav, please add a description on how to obtain the example fluxes files (link to the GitHub download or the `curl` example should be sufficient). I appear to have intuitively cloned the repo and missed this issue.
- [ ] I would also like to see you take another pass at clearly stating the statements of need, specifically `sntools` intended audience. This should be in the JOSS paper, and the documentation.

I think these are small issues and should lead to an acceptance shortly. Great work.
