### 2021-01-11 #2877 @bradkav
Hi @JostMigenda! Thanks for making the latest changes. With https://github.com/JostMigenda/sntools/pull/20 and the latest updates to the readme/documentation, I don't think I have any further comments. Everything seems to run nicely now!

Once the remaining comments raised by @benjaminrose have been addressed, I am very happy to recommend this for publication!
