### 2021-01-11 ca1f22b
Create publish.yml
As discussed in #19
