+++ sntools/__init__.py
@@ -13,7 +13,7 @@ For more extensive documentation, to report issues or to contribute code,
 see https://github.com/JostMigenda/sntools.
 """
 
-__version__ = '0.7.0'
+__version__ = '0.7.1'
 
 
 def setup():
