+++ doc/documentation.tex
@@ -35,7 +35,7 @@
 \newcommand{\nuxbar}{\ensuremath{\bar{\nu}_x}\xspace}
 
 
-\title{Documentation for sntools 0.7\footnote{See \url{https://github.com/JostMigenda/sntools} for the most recent version.}}
+\title{Documentation for sntools 0.7.1\footnote{See \url{https://github.com/JostMigenda/sntools} for the most recent version.}}
 \author{Jost Migenda\footnote{King’s College London, \url{jost.migenda@kcl.ac.uk}}}
 \date{\today}
 
