### 2021-01-11 7fbdc6c
DOC: edit Getting Support and Contributing section
