### 2021-01-11 JostMigenda/sntools#19 @JostMigenda
I’ve [added an action](https://github.com/JostMigenda/sntools/commit/ca1f22b1390aeb06e1c37169d9aa87f340667208) to package up any new release and publish it to PyPI. Made a new release (v0.7.1) a few hours ago to test this and it appears to work well. This action will largely eliminate the possibility of human error & will work without modification for typical changes to sntools (i.e. new detectors, input formats, etc.).

I’ve made a private note to revisit this if there are major changes to the structure of sntools & will also leave this issue open for visibility.
