### 2021-01-11 #2877 @JostMigenda
@bradkav Great; many thanks for your comments and suggestions!

@benjaminrose Quick replies to the three remaining comments:
* I’ve added a sentence on getting support to the README file a few days ago in response to @bradkav, and [have updated](https://github.com/JostMigenda/sntools/commit/7fbdc6c8771c9d6d7061e84cb26a89fdf983a9c9) the documentation accordingly just now. (See the very start of chapter 3 in [this compiled PDF](https://github.com/openjournals/joss-reviews/files/5798257/documentation.pdf).)
* I added a setup function in JostMigenda/sntools#20, which downloads the sample flux file used in the examples. Can you please try this out (`pip install -U sntools` to get v0.7.1, then `python -c 'import sntools; sntools.setup()'`) and let me know if you have any further suggestions?
* I’ll try to look at the statement of need in the next few days.
