### 2021-01-11 #2877 @benjaminrose
@JostMigenda, I think I was reviewing a slightly out of date version of the paper and project. It appears that all my concerns have been addressed.
