### 2021-01-11 #2877 @benjaminrose
@xuanxu I recommend `sntools` for publication in JOSS. @JostMigenda, great work!
