### 2021-01-12 #2877 @bradkav
@xuanxu I'm also now happy to recommend `sntools` for publication. Well done @JostMigenda!
