+++ doc/documentation.bib
@@ -24,43 +24,6 @@
 	Volume = {80},
 	Year = {2020}}
 
-@article{Armbruster1998,
-	Author = {Armbruster, B and others},
-	Collaboration = {KARMEN},
-	Date-Added = {2020-04-22 15:48:53 +0200},
-	Date-Modified = {2020-04-22 15:48:54 +0200},
-	Doi = {10.1016/S0370-2693(98)00087-2},
-	Journal = {Phys. Lett. B},
-	Pages = {15--20},
-	Title = {{Measurement of the weak neutral current excitation C-12(nu(mu) nu'(mu))C*-12(1+,1,15.1-MeV) at E(nu(mu)) = 29.8-MeV}},
-	Volume = {423},
-	Year = {1998}}
-
-@article{Donnelly1979,
-	Author = {Donnelly, T.W. and Peccei, R.D.},
-	Date-Added = {2020-04-22 15:47:21 +0200},
-	Date-Modified = {2020-04-22 15:47:22 +0200},
-	Doi = {10.1016/0370-1573(79)90010-3},
-	Journal = {Phys. Rept.},
-	Pages = {1},
-	Reportnumber = {ITP-597-STANFORD},
-	Title = {{Neutral Current Effects in Nuclei}},
-	Volume = {50},
-	Year = {1979}}
-
-@article{Kolbe1999,
-	Archiveprefix = {arXiv},
-	Author = {Kolbe, E. and Langanke, K. and Vogel, P.},
-	Date-Added = {2020-04-22 15:04:17 +0200},
-	Date-Modified = {2020-04-22 15:04:18 +0200},
-	Doi = {10.1016/S0375-9474(99)00152-9},
-	Eprint = {nucl-th/9903022},
-	Journal = {Nucl. Phys. A},
-	Pages = {91--100},
-	Title = {{Weak reactions on C-12 within the continuum random phase approximation with partial occupancies}},
-	Volume = {652},
-	Year = {1999}}
-
 @article{Askins2015,
 	Archiveprefix = {arXiv},
 	Author = {Askins, M. and Bergevin, M. and Bernstein, A. and Dazeley, S. and Dye, S. T. and Handler, T. and Hatzikoutelis, A. and Hellfeld, D. and Jaffke, P. and Kamyshkov, Y. and Land, B. and Learned, J. G. and Marleau, P. and Mauger, C. and Orebi Gann, G. D. and Roecker, C. and Rountree, S. D. and Shokair, T. M. and Smy, M. and Svoboda, R. and Sweany, M. and Vagins, M. and {van Bibber}, K. A. and Vogelaar, R. B. and Wetstein, M. J. and Yeh, M.},
@@ -74,430 +37,6 @@
 	Title = {{The Physics and Nuclear Nonproliferation Goals of WATCHMAN: A WAter CHerenkov Monitor for ANtineutrinos}},
 	Year = {2015}}
 
-@article{Kolbe2002,
-	Author = {Kolbe, E. and Langanke, K. and Vogel, P.},
-	Date-Added = {2020-04-22 14:06:37 +0200},
-	Date-Modified = {2020-04-22 14:06:37 +0200},
-	Doi = {10.1103/PhysRevD.66.013007},
-	Journal = {Phys. Rev.},
-	Pages = {013007},
-	Slaccitation = {%%CITATION = PHRVA,D66,013007;%%},
-	Title = {{Estimates of weak and electromagnetic nuclear decay signatures for neutrino reactions in Super-Kamiokande}},
-	Volume = {D66},
-	Year = {2002}}
-
-@article{Seadrow2018,
-	Archiveprefix = {arXiv},
[…]
