### 2021-01-12 #2877 @xuanxu
@JostMigenda Whedon reports missing and invalid DOIs for some references but I think they are not used in the paper and can be removed from the `.bib` file so that should not be an issue.
