### 2021-01-12 #2877 @xuanxu
@JostMigenda, everything looks good, you can take a final look at the [pdf draft](https://raw.githubusercontent.com/openjournals/joss-papers/joss.02877/joss.02877/10.21105.joss.02877.pdf) and if you find it OK, here are the next steps:

- Please release a new tagged version from the current master so it includes all the changes made during the review process
- Then archive that latest release in Zenodo (or a similar archival service)
- Check the Zenodo deposit has the correct metadata: title and author names should match the paper; you may also add their ORCID info.

Once you do that please report here the version number and archive DOI
