### 2021-01-12 b570ac8
fix typo in setup function
