### 2021-01-12 2f4c988
added/fixed DOIs Thanks to @whedon https://github.com/openjournals/joss-reviews/issues/2877#issuecomment-758589449
