+++ doc/documentation.bib
@@ -107,6 +107,7 @@
 	Author = {Irene Tamborra and B. Müller and L. Hüdepohl and H.-Th. Janka and Georg G. Raffelt},
 	Date-Added = {2020-04-22 13:47:50 +0200},
 	Date-Modified = {2020-04-22 13:59:54 +0200},
+	Doi = {10.1103/PhysRevD.86.125031},
 	Journal = {Phys. Rev.},
 	Number = {125031},
 	Title = {High-resolution supernova neutrino spectra represented by a simple fit},
@@ -118,6 +119,7 @@
 	Author = {M. T. Keil and Georg G. Raffelt and H.-Th. Janka},
 	Date-Added = {2020-04-22 13:47:44 +0200},
 	Date-Modified = {2020-04-22 13:59:54 +0200},
+	Doi = {10.1086/375130},
 	Journal = {Astrophys. J.},
 	Pages = {971--991},
 	Title = {Monte Carlo Study of Supernova Neutrino Spectra Formation},
@@ -129,6 +131,7 @@
 	Author = {A. Mirizzi and Georg G. Raffelt and P. D. Serpico},
 	Date-Added = {2020-04-22 13:47:29 +0200},
 	Date-Modified = {2020-04-22 13:59:54 +0200},
+	Doi = {10.1088/1475-7516/2006/05/012},
 	Journal = {JCAP},
 	Number = {012},
 	Title = {Earth matter effects in supernova neutrinos: optimal detector locations},
@@ -139,6 +142,7 @@
 	Author = {Amol S. Dighe and Mathias Th. Keil and Georg G. Raffelt},
 	Date-Added = {2020-04-22 13:47:19 +0200},
 	Date-Modified = {2020-04-22 13:59:54 +0200},
+	Doi = {10.1088/1475-7516/2003/06/006},
 	Journal = {JCAP},
 	Number = {06},
 	Pages = {006},
@@ -214,7 +218,7 @@
 	Author = {Raffelt, Georg G. and Smirnov, Alexei {\relax Yu}.},
 	Date-Added = {2020-04-22 13:45:26 +0200},
 	Date-Modified = {2020-04-22 13:59:54 +0200},
-	Doi = {10.1103/PhysRevD.76.081301, 10.1103/PhysRevD.77.029903},
+	Doi = {10.1103/PhysRevD.76.081301},
 	Eprint = {0705.1830},
 	Journal = {Phys. Rev.},
 	Pages = {081301},
@@ -492,6 +496,7 @@
 	Author = {A. Strumia and F. Vissani},
 	Date-Added = {2020-04-22 13:37:39 +0200},
 	Date-Modified = {2020-04-22 13:59:54 +0200},
+	Doi = {10.1016/s0370-2693(03)00616-6},
 	Journal = {Phys. Lett.},
 	Pages = {42--54},
 	Title = {Precise quasielastic neutrino/nucleon cross-section},
@@ -553,7 +558,7 @@
[…]
