### 2021-01-14 #2877 @JostMigenda
Many thanks @benjaminrose and @bradkav for taking the time to review and giving me so many helpful comments!

@xuanxu I have submitted the corresponding physics-focussed paper to AAS yesterday. (@augustfly: manuscript number is AAS29735, in case you need that.) Should I tag & archive the latest release now? Or should we pause this review and do the final steps once the AAS paper is ready?
