### 2021-01-14 #2877 @xuanxu
@JostMigenda I'm going to pause this until the AAS paper has a DOI. After that you can update the `paper.md` file with the AAS journal name and DOI, and then tag & archive a new release.

Please report back here with the AAS paper's DOI, the new version number and the Zenodo (or similar) DOI.

For reference, the DOI of this paper (when it is published) will be `10.21105/joss.2877`. Probably the AAS folks will want that.
