### 2021-02-26 78e672e
add HyperKDR configuration This is the configuration in the 2018 Design Report (before the September 2019 optimization) and is provided for backward compatibility only.
