+++ doc/documentation.tex
@@ -150,7 +150,7 @@ The detector material doesn’t need to be specified explicitly; it is determine
 Select a detector configuration by using the \texttt{--detector <value>} command line option. The following detector configurations are currently implemented:
 
 \begin{description}
-\item[\texttt{HyperK}] Inner detector of Hyper-Kamiokande~\cite{HKDR2018}. Cylinder with a height of \SI{65.8}{m} and a diameter of \SI{64.8}{m}, filled with water.\footnote{In September 2019, the dimensions were changed slightly compared to the 2018 Design Report.}
+\item[\texttt{HyperK}] Inner detector of Hyper-Kamiokande~\cite{HKDR2018}. Cylinder with a height of \SI{65.8}{m} and a diameter of \SI{64.8}{m}, filled with water.\footnote{In September 2019, the dimensions were changed slightly compared to the 2018 Design Report. The old configuration is available as \texttt{HyperKDR}.}
 \item[\texttt{SuperK}] Inner detector of Super-Kamiokande~\cite{Fukuda2003}. Cylinder with a height of \SI{36.2}{m} and a diameter of \SI{33.7}{m}, filled with water.
 \item[\texttt{WATCHMAN}] WATCHMAN detector~\cite{Askins2015}. Cylinder with a height and diameter of \SI{12.8}{m}, filled with water.
 \item[\texttt{WATCHMAN-LS}] Same as \texttt{WATCHMAN}, but filled with liquid scintillator.
