### 2021-04-12 #2877 @JostMigenda
Just to give a quick update: The physics paper was just accepted by ApJ; I expect we’ll hear from their editorial office soon and can then progress here.
