+++ doc/paper.md
@@ -33,7 +33,7 @@ bibliography: documentation.bib
 
 # Optional fields if submitting to a AAS journal too, see this blog post:
 # https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
-aas-doi: 10.3847/xxxxx # Update this with the DOI from AAS once you know it.
+aas-doi: 10.3847/1538-4357/abf7c4 # Update this with the DOI from AAS once you know it.
 aas-journal: Astrophysical Journal # The name of the AAS journal.
 ---
 
