### 2021-04-14 #2877 @arfon
@JostMigenda - I've just head from the AAS that the DOI for their paper is `10.3847/1538-4357/abf7c4`. I've opened a PR to add this to your paper (https://github.com/JostMigenda/sntools/pull/24), but I'm not sure if the journal name also needs updating (i.e., is _Astrophysical Journal_ the correct journal name?)
