### 2021-04-14 #2877 @JostMigenda
@arfon Merged your PR, thanks! ApJ is still correct; though I’ve noticed that the official name is "_The_ Astrophysical Journal"—I’ve updated that.

I’ve just tagged a new release (v0.7.3) and archived it on Zenodo ([10.5281/zenodo.4688040](https://dx.doi.org/10.5281/zenodo.4688040)). Please let me know if you need anything else!
